Page fault handling fix.
Factored out mapping of the physical page as the final generic code
after all fault-specific handling is done.

Fixed the error that zero page didn't have an owner (devzero).

Fixed the error that struct dirent did not have the record length
field as u16 as expected by userspace.

+++ tasks/fs0/include/fs.h
@@ -107,7 +107,7 @@ struct dirbuf {
 struct dirent {
 	u32 inum;			/* Inode number */
 	u32 offset;			/* Dentry offset in its buffer */
-	u32 rlength;			/* Record length */
+	u16 rlength;			/* Record length */
 	u8  name[DIRENT_NAME_MAX];	/* Name string */
 };
